+++ app/components/modal/types/user-profile/index.ts
@@ -1,11 +1,11 @@
 import { action } from '@ember/object';
 import { service } from '@ember/service';
 import Component from '@glimmer/component';
-import { User } from 'potber-client/services/api/types';
+import { Users } from 'potber-client/services/api/types';
 import ModalService from 'potber-client/services/modal';
 
 export interface UserProfileModalOptions {
-  user: User;
+  user: Users.Read;
 }
 
 interface Signature {