+++ app/components/nav/routes/thread/index.ts
@@ -4,17 +4,18 @@ import { service } from '@ember/service';
 import RendererService from 'potber-client/services/renderer';
 import ModalService from 'potber-client/services/modal';
 import RouterService from '@ember/routing/router-service';
-import Thread from 'potber-client/models/thread';
 import CustomSession from 'potber-client/services/custom-session';
 import { appConfig } from 'potber-client/config/app.config';
 import MessagesService from 'potber-client/services/messages';
 import ApiService from 'potber-client/services/api';
 import { IntlService } from 'ember-intl';
 import { trackedFunction } from 'ember-resources/util/function';
+import { ThreadResource } from 'potber-client/routes/authenticated/thread';
 
 export interface Signature {
   Args: {
-    thread: Thread;
+    threadId: string;
+    threadResource: ThreadResource;
   };
 }
 
@@ -28,16 +29,24 @@ export default class NavRoutesThreadComponent extends Component<Signature> {
   @service declare intl: IntlService;
   declare args: Signature['Args'];
 
+  get thread() {
+    return this.args.threadResource.value;
+  }
+
+  get title() {
+    return this.thread?.title;
+  }
+
   get subtitle() {
-    return `Seite ${this.currentPage} von ${this.args.thread.pagesCount}`;
+    return `Seite ${this.currentPage} von ${this.thread?.pagesCount}`;
   }
 
   get nextPageVisible() {
-    return this.currentPage < this.args.thread.pagesCount;
+    return this.thread && this.currentPage < this.thread.pagesCount;
   }
 
   get currentPage() {
-    return this.args.thread.page?.number || 1;
+    return this.thread?.page?.number ?? 1;
   }
 
   get nextPage() {
@@ -52,8 +61,12 @@ export default class NavRoutesThreadComponent extends Component<Signature> {
     return this.currentPage - 1;
   }
 
+  get lastPage() {
+    return this.thread?.pagesCount;
+  }
+
   get originalUrl() {
-    return `${appConfig.forumUrl}thread.php?TID=${this.args.thread.id}`;
+    return `${appConfig.forumUrl}thread.php?TID=${this.thread?.id}`;
   }
 
   get authenticated() {
@@ -61,7 +74,7 @@ export default class NavRoutesThreadComponent extends Component<Signature> {
   }
 
   bookmarkResource = trackedFunction(this, () =>
-    this.api.findBookmarkByThreadId(this.args.thread.id),
+    this.api.findBookmarkByThreadId(this.args.threadId),
   );
 
   get bookmark() {
@@ -94,19 +107,19 @@ export default class NavRoutesThreadComponent extends Component<Signature> {
   handleGoToPage = () => {
     this.modal.input({
       title: 'Gehe zu Seite...',
-      text: `Gib eine Seite zwischen 1 und ${this.args.thread.pagesCount} ein.`,
+      text: `Gib eine Seite zwischen 1 und ${this.thread?.pagesCount} ein.`,
       icon: 'arrow-right',
       label: `Seite`,
       type: 'number',
       minLength: 1,
       maxLength: 5,
       min: 1,
-      max: this.args.thread.pagesCount,
+      max: this.thread?.pagesCount,
       submitLabel: 'Los',
       onSubmit: (value) => {
         this.router.transitionTo('authenticated.thread', {
           queryParams: {
-            TID: this.args.thread.id,
+            TID: this.args.threadId,
             page: value,
           },
         });