+++ app/services/api/types/index.ts
@@ -1,4 +1,3 @@
 export * as Posts from './posts';
-export { Read as Post } from './posts';
 export * as Users from './users';
-export { Read as User } from './users';
+export * as Threads from './threads';