+++ types/ember-truth-helpers/index.d.ts
@@ -1,1 +1,2 @@
 declare module 'ember-truth-helpers/helpers/eq';
+declare module 'ember-truth-helpers/helpers/or';